+++ sdk/core/core-util/src/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export { delay, type DelayOptions } from "./delay.js";
+export { delay, type DelayOptions, calculateExponentialDelayInterval } from "./delay.js";
 export {
   type AbortOptions,
   cancelablePromiseRace,